Working on fixing merge mistakes and preventing regressions
